rpi-spotify-matrix-display/requirements.txt
pillow
numpy
requests
spotipy
RGBMatrixEmulator
